Update filter by deadline function
After filtering, the current function displays a list of tasks with deadline that must be equal to the input deadline. However, this leaves out tasks with deadlines that come before the input deadline.

This change allows the user to see a list of tasks with deadlines specific to and before the input deadline for greater clarity.
